Restructure components and containers
