+++ frontend/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import io from 'socket.io-client';
 
-import ChatContainer from './components/containers/ChatContainer';
+import ChatContainer from './containers/ChatContainer';
 
 const socket = io();
